surf_kat/requirements.txt
django>=1.7,<=1.8
celery>=3.1,<=3.2
docker-py
psycopg2
django-debug-toolbar
pytz
uwsgi
flower
django-celery
aplpy
matplotlib